third_party/WebKit/LayoutTests/http/tests/css/font-face-src-cached.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" href="resources/font-face-via-https.css">
</head>
<body>
<p>Outer page text</p>
<iframe></iframe>
<script>
async_test(function(t) {
  window.addEventListener('load', t.step_func(function() {
    if (window.internals) {
      // Check that the font face is in the Resource::Cached state.
      assert_true(internals.isLoadingFromMemoryCache(
          'https://localhost:8443/css/resources/cors-ahem.php'))
    }

    var iframe = document.querySelector('iframe');

    // This will cause a renderer crash if DCHECKs are enabled and the request
    // for restoring the font face has SSL errors and doesn't have the
    // WebURLRequest::RequestContextFont context (crbug.com/624474).
    iframe.src = 'resources/font-face-src-cached-iframe.html';
    iframe.addEventListener('load', t.step_func(function() {
      t.done();
    }));
  }));
}, 'Test using cached CSS font face src value resources (see ' +
    'CSSFontFaceSrcValue::restoreCachedResourceIfNeeded)');
</script>
</body>
</html>
