third_party/blink/web_tests/platform/win11-arm64/http/tests/css/font-face-src-cached-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: https://localhost:8443/css/resources/cors-ahem.php
CONSOLE WARNING: OTS parsing error: invalid sfntVersion: 1013084704
This is a testharness.js-based test.
FAIL Test using cached CSS font face src value resources (see CSSFontFaceSrcValue::restoreCachedResourceIfNeeded) assert_true: expected true got false
Harness: the test ran to completion.
